+++ time_current-time.go
@@ -9,6 +9,7 @@ import (
 	"time"
 )
 
+//counterfeiter:generate -o mocks/current-time-getter.go --fake-name CurrentTimeGetter . CurrentTimeGetter
 type CurrentTimeGetter interface {
 	Now() time.Time
 }
@@ -19,11 +20,12 @@ func (c CurrentTimeGetterFunc) Now() DateTime {
 	return c()
 }
 
+//counterfeiter:generate -o mocks/current-time-setter.go --fake-name CurrentTimeSetter . CurrentTimeSetter
 type CurrentTimeSetter interface {
 	SetNow(now time.Time)
 }
 
-//counterfeiter:generate -o mocks/time-current-time.go --fake-name TimeCurrentTime . CurrentTime
+//counterfeiter:generate -o mocks/current-time.go --fake-name CurrentTime . CurrentTime
 type CurrentTime interface {
 	CurrentTimeGetter
 	CurrentTimeSetter